data_collection/pyproject.toml
[project]
name = "data-collection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.4",
    "kafka>=1.3.5",
    "kafka-python>=2.2.15",
    "loguru>=0.7.3",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-mypy>=1.0.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "scrapy>=2.13.3",
]
[tool.pytest.ini_options]
pythonpath = [
  "."
]
